feat: /api/user/commandes: en post permet de passer une commande

+++ src/main/java/com/example/quickdinner/model/ProduitCommanderId.java
@@ -13,6 +13,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class ProduitCommanderId implements Serializable {
 
-    private Long commande;
-    private Long produit;
+    private int commande;
+    private int produit;
 }
